backend/requirements.txt
numpy==1.26.4
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.7.4
pydantic-settings==2.9.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
asyncpg==0.29.0
chromadb==0.6.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tenacity==8.2.3
alembic==1.13.0 
openai-whisper==20231117
torch==2.2.0
yt-dlp==2024.3.10
langchain>=0.3
langchain-openai>=0.3
langchain-community>=0.3
openai==1.68.2
tiktoken==0.7.0